Added initial rearrangement of source files due to more test cases to be implemented.
